requirements.txt
pandas==1.2.3
requests==2.25.1
python-decouple==3.4
